Added cases for SPI1 MOSI, MISO, and SPISCK.
Got rid of redundant EVT::core::IO.

Changed the SPI mode in the configSPI call to an enum to minimize user errors.

Changed order in configSPI to be a bool as that is how it is used. minimizes chances of miss input.

Updated f3 to also work with the new changes. All of these will be tested to confirm they still work on thurs.